Filename changed to config/projects.example.yml #6
* Filename changed to config/projects.example.yml

This file will be copied into the container. If name is changed to config/projects.yml, it will be not found during the build of the container.

* Filename changed to config/projects.example.yml

This file will be copied into the container. If name is changed to config/projects.yml, it will be not found during the build of the container.